split enc.cc into six files
enc.cc had grown to 2573 lines doing six unrelated jobs. Every cut lands
on a '////' banner that was already there, so no function body is touched
and no line is rewritten:

  quantize.cc   quant matrices, QuantizeBlock x3, the trellis
  encoders.cc   the per-colorspace Encoder sub-classes and their factories
  entropy.cc    Huffman tables, DC codes, CodeBlock, entropy statistics
  histogram.cc  StoreHisto x3, AnalyseHisto, CollectHistograms
  api.cc        the plain-C entry points, EncoderParam, Encode() overloads
  enc.cc        lifecycle, setters, buffers, dispatch, scan drivers

Cheap because Encoder is already fully declared in sjpegi.h. Six things
had to become visible across the new boundaries; three others stop
travelling and move next to their only reader. Measured at 0%.
